Made it possible to use LoganSquare with multiple modules
This version added an option `jsonMapperLoaderSuffix` to LoganSquare, which allows user to specify suffix to JsonMapperLoader class implementation, user can now name JsonMapperLoader differently, in order to avoid clash on compile time. The behavior will be exactly the same as previous when no suffix specified.

To load multiple JsonMapperLoader, user needs to register their implementation in META-INF/services/com.bluelinelabs.logansquare.internal.JsonMapperLoader.

This is kind of complicated indeed, but it's so far the only compatible way for both Android and Java Standard to use LoganSquare in multiple modules at this time.

 I wish you could consider merge this.